RF | Create package Job and JobCongig, move in them corresponding classes
