+++ src/test/java/school/redrover/FolderTest.java
@@ -4,6 +4,9 @@ import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import school.redrover.model.*;
+import school.redrover.model.Jobs.*;
+import school.redrover.model.JobsConfig.FolderConfigPage;
+import school.redrover.model.JobsConfig.FreestyleProjectConfigPage;
 import school.redrover.model.base.BaseConfigPage;
 import school.redrover.model.base.BaseJobPage;
 import school.redrover.runner.BaseTest;
