+++ src/test/java/school/redrover/MultiConfigurationProjectTest.java
@@ -6,6 +6,8 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.model.*;
+import school.redrover.model.Jobs.MultiConfigurationProjectPage;
+import school.redrover.model.JobsConfig.MultiConfigurationProjectConfigPage;
 import school.redrover.runner.BaseTest;
 import school.redrover.runner.TestUtils;
