+++ src/test/java/school/redrover/ViewsTest.java
@@ -5,6 +5,10 @@ import org.openqa.selenium.By;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.model.*;
+import school.redrover.model.Jobs.FolderPage;
+import school.redrover.model.Jobs.FreestyleProjectPage;
+import school.redrover.model.JobsConfig.FolderConfigPage;
+import school.redrover.model.JobsConfig.FreestyleProjectConfigPage;
 import school.redrover.runner.BaseTest;
 import school.redrover.runner.TestUtils;
