+++ src/test/java/school/redrover/model/Jobs/FolderPage.java
@@ -1,13 +1,13 @@
-package school.redrover.model;
+package school.redrover.model.Jobs;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import school.redrover.model.base.BaseConfigPage;
+import school.redrover.model.*;
+import school.redrover.model.JobsConfig.FolderConfigPage;
 import school.redrover.model.base.BaseJobPage;
-import school.redrover.model.base.BaseMainHeaderPage;
 import school.redrover.runner.TestUtils;
 
 import java.util.List;
