+++ src/test/java/school/redrover/model/Jobs/MultiConfigurationProjectPage.java
@@ -1,9 +1,10 @@
-package school.redrover.model;
+package school.redrover.model.Jobs;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import school.redrover.model.JobsConfig.MultiConfigurationProjectConfigPage;
 import school.redrover.model.base.BaseProjectPage;
 
 public class MultiConfigurationProjectPage extends BaseProjectPage<MultiConfigurationProjectPage> {
