+++ src/test/java/school/redrover/model/JobsConfig/MultiConfigurationProjectConfigPage.java
@@ -1,7 +1,8 @@
-package school.redrover.model;
+package school.redrover.model.JobsConfig;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
+import school.redrover.model.Jobs.MultiConfigurationProjectPage;
 import school.redrover.model.base.BaseConfigProjectsPage;
 
 public class MultiConfigurationProjectConfigPage extends BaseConfigProjectsPage<MultiConfigurationProjectConfigPage, MultiConfigurationProjectPage> {
