+++ src/test/java/school/redrover/model/JobsConfig/MultibranchPipelineConfigPage.java
@@ -1,8 +1,9 @@
-package school.redrover.model;
+package school.redrover.model.JobsConfig;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
+import school.redrover.model.Jobs.MultibranchPipelinePage;
 import school.redrover.model.base.BaseConfigFoldersPage;
 
 public class MultibranchPipelineConfigPage extends BaseConfigFoldersPage<MultibranchPipelineConfigPage, MultibranchPipelinePage> {
