+++ src/test/java/school/redrover/model/JobsConfig/PipelineConfigPage.java
@@ -1,11 +1,11 @@
-package school.redrover.model;
+package school.redrover.model.JobsConfig;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
-import school.redrover.model.base.BaseConfigPage;
+import school.redrover.model.Jobs.PipelinePage;
 import school.redrover.model.base.BaseConfigProjectsPage;
 import school.redrover.runner.TestUtils;
