+++ src/test/java/school/redrover/model/ViewFolderPage.java
@@ -2,8 +2,8 @@ package school.redrover.model;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import school.redrover.model.Jobs.FolderPage;
 import school.redrover.model.base.BaseMainHeaderPage;
 
 public class ViewFolderPage extends BaseMainHeaderPage<ViewFolderPage> {
