+++ src/test/java/school/redrover/model/base/BaseConfigFoldersPage.java
@@ -3,8 +3,8 @@ package school.redrover.model.base;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import school.redrover.model.MultibranchPipelineConfigPage;
-import school.redrover.model.MultibranchPipelinePage;
+import school.redrover.model.JobsConfig.MultibranchPipelineConfigPage;
+import school.redrover.model.Jobs.MultibranchPipelinePage;
 import school.redrover.runner.TestUtils;
 
 public abstract class BaseConfigFoldersPage<Self extends BaseConfigPage<?, ?>, FolderPage extends BaseMainHeaderPage<?>> extends BaseConfigPage<Self, FolderPage>{
