+++ src/test/java/school/redrover/model/base/BaseConfigPage.java
@@ -1,9 +1,7 @@
 package school.redrover.model.base;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import school.redrover.model.PipelineConfigPage;
 
 public abstract class BaseConfigPage<Self extends BaseConfigPage<?, ?>, JobPage extends BaseMainHeaderPage<?>> extends BaseMainHeaderPage<Self> {
