+++ src/test/java/school/redrover/runner/TestUtils.java
@@ -7,6 +7,8 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import school.redrover.model.*;
+import school.redrover.model.Jobs.*;
+import school.redrover.model.JobsConfig.*;
 import school.redrover.model.base.BaseConfigPage;
 import school.redrover.model.base.BaseModel;
